removed unused libraries from build path, fixed spelling, removed not needed scala files
