+++ SparkStreaming/.classpath
@@ -169,7 +169,6 @@
 	<classpathentry kind="lib" path="lib/pyrolite-4.9.jar"/>
 	<classpathentry kind="lib" path="lib/RoaringBitmap-0.5.11.jar"/>
 	<classpathentry kind="lib" path="lib/scala-compiler-2.11.8.jar"/>
-	<classpathentry kind="lib" path="lib/scala-library-2.11.8.jar"/>
 	<classpathentry kind="lib" path="lib/scala-parser-combinators_2.11-1.0.4.jar"/>
 	<classpathentry kind="lib" path="lib/scala-reflect-2.11.8.jar"/>
 	<classpathentry kind="lib" path="lib/scala-xml_2.11-1.0.2.jar"/>
@@ -193,8 +192,6 @@
 	<classpathentry kind="lib" path="lib/spark-sketch_2.11-2.0.1.jar"/>
 	<classpathentry kind="lib" path="lib/spark-sql_2.11-2.0.1.jar"/>
 	<classpathentry kind="lib" path="lib/spark-streaming_2.11-2.0.1.jar"/>
-	<classpathentry kind="lib" path="lib/spark-streaming-twitter_2.10-1.6.3.jar"/>
-	<classpathentry kind="lib" path="lib/spark-streaming-twitter_2.11-1.6.2 2.jar"/>
 	<classpathentry kind="lib" path="lib/spark-streaming-twitter_2.11-1.6.2.jar"/>
 	<classpathentry kind="lib" path="lib/spark-tags_2.11-2.0.1.jar"/>
 	<classpathentry kind="lib" path="lib/spark-unsafe_2.11-2.0.1.jar"/>
