--- SparkStreaming/src/Helloworld.scala
@@ -1,5 +1,0 @@
-object Helloworld {
-  def main(args: Array[String]){
-    println("Hello, world!")
-  }
-}
